requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-Login
Flask-WTF
email-validator
Pillow
